refactor(app): 将main函数重命名为Start并创建新的入口文件
- 将app包中的main函数重命名为Start函数
- 创建新的main.go文件作为应用入口点
- 更新包声明从main到app以避免冲突
- 移除原始main.go文件中的重复代码
- 通过导入app包并调用Start函数来启动应用
- 确保代码结构更清晰便于测试和维护
