+++ app/wire_gen.go
@@ -4,7 +4,7 @@
 //go:build !wireinject
 // +build !wireinject
 
-package main
+package app
 
 import (
 	"go-zero-box/app/internal/command"
