+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "go-zero-box/app"
+
+func main() {
+	app.Start()
+}
